requirements.txt
aiogram==3.4.1
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
pytz==2024.1
aiohttp>=3.9.3,<4
APScheduler==3.10.4
python-dotenv==1.0.1